requirements.txt
numpy
flask
gunicorn
flask-cors
pillow
torch==1.12.0+cpu
torchvision==0.13.0+cpu

--extra-index-url https://download.pytorch.org/whl/cpu
